git-svn-id: http://svn.code.sf.net/p/chapel/code/branches/collaborations/caseyb@20082 3a8e244f-b0f2-452b-bcba-4c88e055c3ca

+++ runtime/include/bulkget.h
@@ -5,11 +5,9 @@
 #include "qbuffer.h"
 #include "qio_style.h"
 
-qbytes_t* bulk_get_bytes(int32_t src_locale, qbytes_t* src_addr);
+qbytes_t* bulk_get_bytes(int64_t src_locale, qbytes_t* src_addr);
 
-//void bulk_get_style(int32_t src_local, qio_style_t* dst_addr, qio_style_t* src_addr);
-
-err_t bulk_put_buffer(int32_t dst_locale, void* dst_addr, int64_t dst_len,
+err_t bulk_put_buffer(int64_t dst_locale, void* dst_addr, int64_t dst_len,
                       qbuffer_t* buf, qbuffer_iter_t start, qbuffer_iter_t end);
