+++ runtime/include/chpl-comm-locales.h
@@ -12,14 +12,14 @@
 // routine returns a value, that value needs to be consistent across
 // multiple calls to the routine.
 //
-int chpl_comm_default_num_locales(void);
+int64_t chpl_comm_default_num_locales(void);
 
 //
 // This routine allows a comm layer to screen the number of locales to
 // be used.  In particular, if a number exceeding some sort of maximum
 // was provided, an error should be reported.
 //
-void chpl_comm_verify_num_locales(int32_t proposedNumLocales);
+void chpl_comm_verify_num_locales(int64_t proposedNumLocales);
 
 
 #endif
