+++ runtime/src/comm/armci/comm-armci-locales.c
@@ -12,7 +12,7 @@
 // routine returns a value, that value needs to be consistent across
 // multiple calls to the routine.
 //
-int chpl_comm_default_num_locales(void) {
+int64_t chpl_comm_default_num_locales(void) {
   return chpl_specify_locales_error();
 }
 
@@ -22,7 +22,7 @@ int chpl_comm_default_num_locales(void) {
 // be used.  In particular, if a number exceeding some sort of maximum
 // was provided, an error should be reported.
 //
-void chpl_comm_verify_num_locales(int32_t proposedNumLocales) {
+void chpl_comm_verify_num_locales(int64_t proposedNumLocales) {
   // This can probably remain empty for ARMCI unless there is any sort of
   // maximum that you'd need to check against
 }
