+++ runtime/src/comm/none/comm-none-locales.c
@@ -2,11 +2,11 @@
 #include "chpl-comm-locales.h"
 #include "error.h"
 
-int chpl_comm_default_num_locales(void) {
+int64_t chpl_comm_default_num_locales(void) {
   return 1;
 }
 
-void chpl_comm_verify_num_locales(int32_t proposedNumLocales) {
+void chpl_comm_verify_num_locales(int64_t proposedNumLocales) {
   if (proposedNumLocales != 1) {
     chpl_error("Only 1 locale may be used for CHPL_COMM layer 'none'\n"
                "To use multiple locales, see $CHPL_HOME/doc/README.multilocale",
