+++ test/extern/sungeun/array_ops.c
@@ -2,13 +2,13 @@
 #include "array_ops.h"
 
 void
-clear_int_array(int *A, int len) {
-  memset(A, 0, len*sizeof(int));
+clear_int_array(int64_t *A, int len) {
+  memset(A, 0, len*sizeof(int64_t));
 }
 
 
 void
-copy_int_array(int *dest, int *src, int len) {
-  memcpy(dest, src, len*sizeof(int));
+copy_int_array(int64_t *dest, int64_t *src, int len) {
+  memcpy(dest, src, len*sizeof(int64_t));
 }
